Code refactoring

#167 Move components into related page components optimizing code organization and cohesion
