+++ src/app/pages/add-new-property/add-condo-dialog/add-condo-dialog.component.scss
