+++ src/app/pages/add-new-property/add-locker-dialog/add-locker-dialog.component.scss
