+++ src/app/pages/add-new-property/add-parking-dialog/add-parking-dialog.component.scss
