+++ src/app/pages/building-info/building-overview/building-overview.component.scss
