+++ src/app/pages/building-info/condo/condo.component.html
