+++ src/app/pages/building-info/condo/condo.component.spec.ts
