+++ src/app/pages/building-info/condo/condo.component.ts
