+++ src/app/pages/building-info/locker/locker.component.html
