+++ src/app/pages/building-info/locker/locker.component.spec.ts
