+++ src/app/pages/building-info/locker/locker.component.ts
