+++ src/app/pages/building-info/parking-spot/parking-spot.component.spec.ts
