+++ src/app/pages/key-registration/key-registration.component.html
